move ui files to unified structure
