+++ apps/web/app/(app)/environments/[environmentId]/(actionsAndAttributes)/actions/components/ActionSettingsTab.tsx
@@ -13,9 +13,7 @@ import { toast } from "react-hot-toast";
 import { getAccessFlags } from "@formbricks/lib/membership/utils";
 import { TActionClassInput, TActionClassNoCodeConfig, TNoCodeConfig } from "@formbricks/types/actionClasses";
 import { TMembershipRole } from "@formbricks/types/memberships";
-import CssSelector from "@formbricks/ui/Actions/CssSelector";
-import InnerHtmlSelector from "@formbricks/ui/Actions/InnerHtmlSelector";
-import PageUrlSelector from "@formbricks/ui/Actions/PageUrlSelector";
+import { CssSelector, InnerHtmlSelector, PageUrlSelector } from "@formbricks/ui/Actions";
 import { Button } from "@formbricks/ui/Button";
 import { DeleteDialog } from "@formbricks/ui/DeleteDialog";
 import { Input } from "@formbricks/ui/Input";
